[INLONG-10881][SDK] Transform SQL support Concat_ws function
- add distinct UT to avoid conflict
- add description of concat_ws function
